Feat: Add UI; add answer-checking feature
It uses a randomizer to randomly select an item from the quiz_items_dict.
Once selected, the item is then printed as a question. After so, the user
will be asked to input an answer. That answer is then compared to the
correct answer before being stored in a dictionary (finished_items).